Documentação das classes do pattern composite
