+++ src/com/examples/composite/Inimigo.java
@@ -1,9 +1,17 @@
 package com.examples.composite;
 
+import com.examples.composite.interfaces.InimigoModel;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Inimigo implements InimigoModel{
+/**
+ * Definição da classe Inimigo, que implementa o método render definido na interface InimigoModel.
+ * Essa classe segue as mesmas definições da classe "Funcionário".
+ * Nessa classe, define-se os métodos e propriedades dos inimigos, e o método "toString()" para formatar
+ * a exibição dos dados do objeto.
+ */
+
+public class Inimigo implements InimigoModel {
 
     public String tituloInimigo = "";
     int tamanho = 0;
@@ -17,17 +25,23 @@ public class Inimigo implements InimigoModel{
         this.alcanceAtaque = alcanceAtaque;
     }
 
+    //Método que adiciona um objeto subordinado a um objeto atual
     public void add(Inimigo i)
     {
         especializacoes.add(i);
     }
 
+    /*Método para remoção de subordinados. Para a simplificação do exemplo
+    * a única validação é se o inimigo indicado existe.*/
     public void remove(Inimigo i)
     {
         if (especializacoes.contains(i))
             especializacoes.remove(i);
     }
 
+   /*Implementação particular de um método "toString()" que obtem os dados do inimigo
+    * e estrutura a impressão.
+    * */
     public String render()
     {
         return "Inimigo ativo: " + this.tituloInimigo +
