--- src/com/examples/composite/InimigoModel.java
@@ -1,10 +1,0 @@
-package com.examples.composite;
-
-/**
- * Created by ricardo on 1/31/17.
- */
-public interface InimigoModel {
-
-    String render();
-
-}
