+++ src/com/examples/composite/interfaces/InimigoModel.java
@@ -1,0 +1,11 @@
+package com.examples.composite.interfaces;
+
+/**
+ * Para e execução do exemplo "Inimigo" criamos uma interface que expõe o método comum a todos os
+ * inimigos. Essa abordagem é opcional para a implementação deste pattern nesta linguagem.
+ */
+public interface InimigoModel {
+
+    String render();
+
+}
