+++ src/com/examples/decorator/BaseWeapon.java
@@ -1,5 +1,7 @@
 package com.examples.decorator;
 
+import com.examples.decorator.interfaces.Weapon;
+
 public class BaseWeapon implements Weapon {
     @Override
     public void load() {
