+++ src/com/examples/decorator/WeaponAccessory.java
@@ -1,5 +1,7 @@
 package com.examples.decorator;
 
+import com.examples.decorator.interfaces.Weapon;
+
 public abstract class WeaponAccessory implements Weapon {
     protected Weapon decoratedWeapon;
